Do not fail if drain node fails (#1455)

+++ templates/drain.go
@@ -3,6 +3,11 @@ package templates
 const drainTpl = `
 NODENAME=$(sudo kubectl get no -o wide|grep {{ .PrivateIP }}| awk '{ print $1 }')
 
+if [ -z $NODENAME ]
+then
+	exit 0
+fi
+
 sudo kubectl drain $NODENAME \
 --ignore-daemonsets --force --delete-local-data
